fix: add.js use isFileAlreadyExist

+++ src/commands/add.js
@@ -2,19 +2,19 @@ import * as fs from 'fs/promises';
 import { currentDirectory } from '../file-manager.mjs';
 import showDirectory from '../utils/showDirectory.js';
 import path from 'path';
+import isFileAlreadyExist from '../utils/isFileAlreadyExist.js';
 
 async function executeAdd(fileName) {
-    // Create empty file in current working directory:
-    const pathToFile = path.join(currentDirectory.path, fileName);
-    try{
-        await fs.access(pathToFile);
-        console.log("The file already exist")
-        console.log("Operation failed")
-    } catch(err){
-        await fs.writeFile(pathToFile, "", {flag:'wx'})
-        console.log('the file is added')
-        showDirectory();
-    }
+  // Create empty file in current working directory:
+  const pathToFile = path.join(currentDirectory.path, fileName);
+  
+  if(await isFileAlreadyExist(pathToFile)){
+      return;
+  };
+
+  await fs.writeFile(pathToFile, "", {flag:'wx'})
+  console.log('the file is added')
+  showDirectory();
 }
 
 export default executeAdd;
